Version 0.0.4:
Added tests for GridFTP, including starting and stopping the server.

Added code to generate a grid-mapfile, moving aside an existing one if
needed.  The original mapfile is restored after testing.

Added the -d command-line option to enable dumping of all commands and
their output; the corresponding log file is removed at the end of the
run in any case.  This is all temporary until we have our own testing
framework.

Improved package uninstalls, especially in the case where both the 32-
and 64-bit versions of a package are installed.

Only attempt to install a package if it does not seem to be installed.


git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@13183 4e558342-562e-0410-864c-e07659590f8c