Add test for cvmfs_config probe for CVMFS 2.1 (SOFTWARE-1097)
Since it probes the oasis.opensciencegrid.org server, I had to add the oasis-config package as a dependency.
oasis-config is not part of osg-tested-internal, so the probe test will be skipped in the nightlies.
(Since it was moved to a separate method though, the rest of the cvmfs tests aren't affected).

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@17120 4e558342-562e-0410-864c-e07659590f8c