Don't require root for tarball tests (SOFTWARE-1002)
This is prep work for running the tarball tests as an HTCondor job.
This required several fixes:

Allow specifying helper files directory (since it may not be in a standard location)
This is done with the 'tarball_helper_files' config option.

Allow specifying location of tarball tests (since they may not be in a standard location)
This is done with the '--tarball-test-dir' command-line option.

Work around core.check_system trying to su to the test user.
If the arg 'user' is True, core.check_system tries to su to the test user,
which doesn't work if osg-test is not running as root.
On the other hand, if 'user' is False, then it doesn't switch to the test user
even if osg-test is running as root. In order to make the tarball tests work in
both cases, I made wrappers around core.check_system so user=True if we're root
and user=False if not.

Also pulled check_osgrun out of TestTarball since there's no reason it has to
be a method.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@17582 4e558342-562e-0410-864c-e07659590f8c