Add tests for GSI-OpenSSH (SOFTWARE-2390)
The tests are one run of gsissh to see if we can connect and run a
simple command, and two runs of gsiscp to see if we can copy files
to/from the remote server.  These tests require gsi-openssh-server,
gsi-openssh-clients, and a proxy.  The code for setting up the file
transfer tests was based off of the gfal2 test code.

A more global change: I changed special_user.py to give the vdttest user
a randomly-generated password.  SSH (including GSISSH) will consider a
user 'disabled' and refuse to log in as that user if the user does not
have a valid encrypted password in /etc/shadow.