allow virtual dependencies for torque packages
Vaguely SOFTWARE-2484 related.

Apparently some (non-epel, oops) flavors of the torque/pbs packages have
a little different layout, and, for instance, 'torque-mom' is actually a
virtual capability provided by the 'torque-client' rpm.  In that case,
we really only care about 'torque-mom' as a dependency, not necessarily
as a package name.